Moved app folder to src
