+++ playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
-  testDir: "./app",
+  testDir: "./src/app",
   testMatch: "*.spec.ts",
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
