+++ src/app/page.spec.ts
@@ -1,17 +1,17 @@
 import { expect } from "@playwright/test";
-import { test } from "../../../src/test/browser/fixtures";
-import { getTabKey } from "../../../src/test/browser/utils";
-import { mockedSearchResponse } from "../../../src/test/api/tenor/mocks/searchResponse";
-import { testLayout } from "../../../src/test/browser/shared";
-import { gifLimit } from "../../../src/constants";
+import { test } from "../test/browser/fixtures";
+import { getTabKey } from "../test/browser/utils";
+import { mockedFeaturedResponse } from "../test/api/tenor/mocks/featuredResponse";
+import { testLayout } from "../test/browser/shared";
+import { gifLimit } from "../constants";
 
 test("initial render", async ({ page, baseUrl }) => {
-  await page.goto(`${baseUrl}/search/dog`);
+  await page.goto(baseUrl);
 
   await testLayout(page);
 
   const searchInput = page.getByRole("textbox", { name: "search" });
-  await expect(searchInput).toHaveValue("dog");
+  await expect(searchInput).toBeEmpty();
 
   const gifs = page.getByRole("button", {
     name: /Load full preview of gif with description/,
@@ -24,7 +24,7 @@ test("selecting a gif and copying the url", async ({
   baseUrl,
   browserName,
 }) => {
-  await page.goto(`${baseUrl}/search/dog`);
+  await page.goto(baseUrl);
 
   const gifs = page.getByRole("button", {
     name: /Load full preview of gif with description/,
@@ -45,16 +45,15 @@ test("selecting a gif and copying the url", async ({
       navigator.clipboard.readText(),
     );
     expect(clipboardText).toContain(
-      mockedSearchResponse.results[0].media_formats.mediumgif.url,
+      mockedFeaturedResponse.results[0].media_formats.mediumgif.url,
     );
   }
 });
 
 test("searching for gifs", async ({ page, baseUrl }) => {
-  await page.goto(`${baseUrl}/search/dog`);
+  await page.goto(baseUrl);
 
   const searchInput = page.getByRole("textbox", { name: "search" });
-  await searchInput.clear();
   await searchInput.fill("cats");
 
   const searchButton = page.getByRole("button", { name: "Search" });
@@ -64,7 +63,7 @@ test("searching for gifs", async ({ page, baseUrl }) => {
 });
 
 test("keyboard navigation of gifs", async ({ page, baseUrl, browserName }) => {
-  await page.goto(`${baseUrl}/search/dog`);
+  await page.goto(baseUrl);
   const tabKey = getTabKey(browserName);
 
   const searchInput = page.getByRole("textbox", { name: "search" });
@@ -111,7 +110,7 @@ test("keyboard navigation of gifs", async ({ page, baseUrl, browserName }) => {
       navigator.clipboard.readText(),
     );
     expect(clipboardText).toContain(
-      mockedSearchResponse.results[1].media_formats.mediumgif.url,
+      mockedFeaturedResponse.results[1].media_formats.mediumgif.url,
     );
   }
 });
@@ -121,12 +120,11 @@ test("keyboard navigation of searching for gifs", async ({
   baseUrl,
   browserName,
 }) => {
-  await page.goto(`${baseUrl}/search/dog`);
+  await page.goto(baseUrl);
   const tabKey = getTabKey(browserName);
 
   const searchInput = page.getByRole("textbox", { name: "search" });
   await expect(searchInput).toBeFocused();
-  await searchInput.clear();
 
   await page.keyboard.insertText("cats");
   await page.keyboard.press(tabKey);
