+++ src/app/search/page.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import { test } from "../../src/test/browser/fixtures";
+import { test } from "../../test/browser/fixtures";
 
 test("redirection to landing page", async ({ page, baseUrl }) => {
   await page.goto(`${baseUrl}/search`);
